+++ src/flet_easy/jwt.py
@@ -1,8 +1,12 @@
+import contextlib
 from secrets import token_bytes
 from typing import Any, Dict, Union
 
-from jwt import DecodeError, ExpiredSignatureError, InvalidKeyError
-from rsa import newkeys
+with contextlib.suppress(ImportError):
+    from jwt import DecodeError, ExpiredSignatureError, InvalidKeyError
+
+with contextlib.suppress(ImportError):
+    from rsa import newkeys
 
 from flet_easy.datasy import Datasy, evaluate_secret_key
 from flet_easy.extra import Msg