+++ tests/flet-app/components/__init__.py
@@ -1,0 +1,5 @@
+from components.counter import Counter
+from components.drawer import Drawer
+from components.login import Login
+from components.register import Register
+from components.custom import Custom