+++ tests/flet-app/components/custom.py
@@ -1,0 +1,31 @@
+import flet as ft
+
+
+class Custom:
+    def __init__():
+        pass
+
+    def custom_appbar(self):
+        return ft.AppBar(
+            title=ft.Text("App - Using a class as a view"),
+            actions=[
+                ft.Row(
+                    controls=[
+                        ft.FilledButton(
+                            "Go back",
+                            on_click=self.data.go_back(),
+                            style=ft.ButtonStyle(
+                                bgcolor=ft.colors.RED_600,
+                            ),
+                        ),
+                        ft.FilledButton(
+                            "Page 2",
+                            on_click=self.data.go("/counter/ts"),
+                            style=ft.ButtonStyle(
+                                bgcolor=ft.colors.AMBER_500,
+                            ),
+                        ),
+                    ]
+                )
+            ],
+        )