+++ tests/flet/components/__init__.py
@@ -1,4 +1,0 @@
-from components.counter import Counter
-from components.drawer import Drawer
-from components.login import Login
-from components.register import Register